fix: fix windows & linux builds

+++ src/memory/process.rs
@@ -9,8 +9,8 @@ use windows::Win32::Foundation::HANDLE;
 
 #[derive(Debug)]
 pub struct MemoryRegion {
-    pub from: i32,
-    pub size: i32
+    pub from: usize,
+    pub size: usize, 
 }
 
 macro_rules! prim_read_impl {
